sample_code/bedrock-chatbot/requirements.txt
streamlit>=1.15
boto3>=1.26
python-dotenv>=1.0